zend_portability: Remove `EMPTY_SWITCH_DEFAULT_CASE()`
This macro is exactly as long as explicitly spelling out its definition
(excluding the useless `break;`), but hides control flow in its definition.
This is giving tools like Coccinelle some troubles to understand the code and
can also be confusing for the human reader, because it appears as if it is a
statement that belongs to the current case rather than opening a new case.