uri: Make uri_parser_rfc3986.h usable for external extensions
This header could not previously be used due to the `uriparser/Uri.h` include,
which is required for the struct definition. Since this struct is considered an
implementation detail, we can just make it opaque, preserving type safety, but
without allowing external users to touch its contents.

Fixes php/php-src#19868.