Clarify what pm.max_spawn_rate means
I was confused by "number of rate", and it looks like this is actually the number of processes to spawn at one time -- so not a rate at all, but obviously we can't change the configuration variable's _name_, but we can improve the explanation! 

I'll submit a matching PR to php/doc-en as well.