.github: Bust the nightly CI cache on Sunday instead of Monday (#20242)
The nightly matrix with an empty cache takes several hours to complete due to
the amount of jobs. This will effectively block the entire CI for the php
organization since there is an organization-wide limit of 20 jobs. Move the
cache buster job to Sunday to make it less likely for folks to fight with the
Nightly build for resources.