Add OpenSSL TLS configurable session resumption support
This adds support for various session options to the stream SSL context.
It allows setting a new session callback and session data on the client,
and get and remove session callbacks on the server. The server also offers
options to configure session cache parameters and the number of session
tickets. A new Openssl\Session class is introduced for session
import/export and introspection, along with Openssl\OpensslException
as the base exception for the extension.

RFC: https://wiki.php.net/rfc/tls_session_resumption_api

Closes GH-20296