Do not pass a CV slot to zend_assign_to_typed_ref_ex() with value_type=IS_TMP_VAR
zend_assign_to_typed_ref_ex() receives a zval slot and tries to dtor it when
value_type=IS_TMP_VAR. This slot may be modified before dtor if it was in fact a
CV slot. In particular it may be turned into a ref, replaced by anyother ref, or
unset; leading to UAFs or leaks.

Fix by passing a stack slot instead. This is consistent with
zend_assign_to_typed_prop().