NEWS: update for PHP 8.5.0RC5, fix next release
Copy the NEWS entries from the RC5 tag with the changes included there. One of
those entries was already listed in the section for the next release from
PHP-8.5; remove it from that section. Also update the header at the top for
what release is next - the next release from the PHP-8.5 branch is going to be
PHP 8.5.1.

[skip ci]