Zend: reengineer how __debugInfo() magic method is handed
The primary motivation is to get rid of the bailout by using a type error, but also properly abort the output when an exception is thrown or the deprecation notice is promoted to an exception.

This is effectively an internal API change as extensions are no longer allowed to return NULL to fast path an empty array.