Fix GH-18956: PHP-FPM Process Count Inconsistencies
This fixes extra increments on keep alive that happen for follow up
request in headers reading stage. It is because the accepting stage
is only done on the first request. It adds a new flag to the on_read
fastcgi callback that sets whether the kept alive request is being
done and then skips the further active increments.

It also fixes issue with incorrect decrement of the active number of
processes. This was done in accepting stage even on the first
which might still be problematic if there are already some running
processes as it decrements their number first and result in incorrect
total (lower than the actual number). The fix is to skip this decrement
of active processes on the first accept.

Closes GH-19191