MSVC: Fix warning C4244 in zend_get_gc_buffer_use
This PR addresses a compilation warning of the form:
warning C4244: '=': conversion from '__int64' to 'int', possible loss of data
The expression gc_buffer->cur - gc_buffer->start produces a temporary of type ptrdiff_t, which is then assigned to an int. It causes said compilation warning if these types do not match.