uri: Update to uriparser-1.0.0
We're now cleanly back on a released version. No functional changes since the
last import of the library.

Version 0.9.10 never existed, the minimum version in config.m4 was already
increased in anticipation of a new release that contained necessary bugfixes to
prevent building against a uriparser without these fixes. It therefore also is
adjusted to 1.0.0 for correctness without having an impact.