Fix GH-20733: heap buffer overflow in optimizer
Some (conditional) jump instructions can be the last one in the
op_array, because they can jump to themselves. In those cases `i + 1` in
the CFG builder can point to outside the op_array because `i` is already
the last opline.
To solve this we need to check against the end and prevent setting the
successor out of bounds.